refactor: reorganize Bank types into separate folder from BankConnection
- Create new splice-api/src/banks/ folder for Bank-related types
- Move Bank interface, BankSourceType enum, and AvailableBankResponse from bank-connections/
- Update bank-connections/types.ts to only contain BankConnection-related types
- Import Bank types from banks folder in bank-connections
- Export banks module from main splice-api index
- Improves code organization by separating Bank and BankConnection concerns
