+++ splice-api/src/banks/index.ts
@@ -1,0 +1,2 @@
+export * from './requests';
+export * from './types';
