+++ splice-api/src/banks/requests.ts
@@ -1,0 +1,2 @@
+// Bank-related request types will be added here as needed
+export {};
