+++ splice-api/src/index.ts
@@ -1,5 +1,6 @@
 export * from './api-key-store';
 export * from './bank-connections';
+export * from './banks';
 export * from './scraper';
 export * from './transactions';
 export * from './users';
